+++ indico/modules/rb/client/js/__tests__/util.test.js
@@ -9,7 +9,7 @@ import moment from 'moment';
 
 import {renderRecurrenceWeekdays} from '../util';
 
-describe('renderRecurrenceWeekdays', () => {
+describe('can render the recurrence weekdays', () => {
   it('should return null if weekdays array is empty', () => {
     const weekdays = [];
     expect(renderRecurrenceWeekdays(weekdays)).toBeNull();
@@ -32,38 +32,38 @@ describe('renderRecurrenceWeekdays', () => {
 
   it('correctly formats a single weekday', () => {
     const formatted = renderRecurrenceWeekdays(['mon']);
-    expect(formatted).toBe('Monday');
+    expect(formatted).toBe('Every Monday');
   });
 
   it('correctly formats two weekdays', () => {
     const formatted = renderRecurrenceWeekdays(['mon', 'tue']);
-    expect(formatted).toBe('Monday and Tuesday');
+    expect(formatted).toBe('Every Monday and Tuesday');
   });
 
   it('correctly formats multiple weekdays', () => {
     const formatted = renderRecurrenceWeekdays(['mon', 'tue', 'wed']);
-    expect(formatted).toBe('Monday, Tuesday, and Wednesday');
+    expect(formatted).toBe('Every Monday, Tuesday, and Wednesday');
   });
 
   it('correctly formats weekdays in non-sequential order', () => {
     const formatted = renderRecurrenceWeekdays(['wed', 'mon', 'fri']);
-    expect(formatted).toBe('Monday, Wednesday, and Friday');
+    expect(formatted).toBe('Every Monday, Wednesday, and Friday');
   });
 
   it('should handle all weekdays', () => {
     const weekdays = ['mon', 'tue', 'wed', 'thu', 'fri', 'sat', 'sun'];
-    const expected = 'Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, and Sunday';
+    const expected = 'Every Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, and Sunday';
     expect(renderRecurrenceWeekdays(weekdays)).toBe(expected);
   });
 
   it('should handle all weekdays in non-sequential order', () => {
     const weekdays = ['wed', 'mon', 'fri', 'tue', 'sun', 'thu', 'sat'];
-    const expected = 'Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, and Sunday';
+    const expected = 'Every Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, and Sunday';
     expect(renderRecurrenceWeekdays(weekdays)).toBe(expected);
   });
 });
 
-describe('renderLocalizedRecurrenceWeekdays', () => {
+describe('can render the recurrence weekdays whilst taking the locale into account', () => {
   const unorderedWeekdays = ['wed', 'mon', 'fri', 'tue', 'sun', 'thu', 'sat'];
   const cases = [
     ['de', 'Montag, Dienstag, Mittwoch, Donnerstag, Freitag, Samstag und Sonntag'],
@@ -80,6 +80,7 @@ describe('renderLocalizedRecurrenceWeekdays', () => {
     ['pl', 'poniedziałek, wtorek, środa, czwartek, piątek, sobota i niedziela'],
     ['mn', 'Даваа, Мягмар, Лхагва, Пүрэв, Баасан, Бямба, Ням'],
     ['uk', 'понеділок, вівторок, середа, четвер, п’ятниця, субота і неділя'],
+    ['hu', 'hétfő, kedd, szerda, csütörtök, péntek, szombat és vasárnap'],
     ['zh-cn', '星期一、星期二、星期三、星期四、星期五、星期六和星期日'],
   ];
 
@@ -89,7 +90,73 @@ describe('renderLocalizedRecurrenceWeekdays', () => {
       const oldLocale = moment.locale();
       moment.locale(locale);
       const formatted = renderRecurrenceWeekdays(unorderedWeekdays);
-      expect(formatted).toBe(expected);
+      expect(formatted).toBe(`Every ${expected}`);
+      moment.locale(oldLocale);
+    }
+  );
+});
+
+describe('can render the recurrence weekdays with support for intervals', () => {
+  it('should handle a single weekday with an interval', () => {
+    const weekdays = ['mon'];
+    const interval = 2;
+    const expected = 'Every 2 weeks on Monday';
+    expect(renderRecurrenceWeekdays(weekdays, interval)).toBe(expected);
+  });
+
+  it('should handle multiple weekdays with an interval', () => {
+    const weekdays = ['mon', 'tue', 'wed'];
+    const interval = 3;
+    const expected = 'Every 3 weeks on Monday, Tuesday, and Wednesday';
+    expect(renderRecurrenceWeekdays(weekdays, interval)).toBe(expected);
+  });
+
+  it('should handle all weekdays with an interval', () => {
+    const weekdays = ['mon', 'tue', 'wed', 'thu', 'fri', 'sat', 'sun'];
+    const interval = 4;
+    const expected =
+      'Every 4 weeks on Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, and Sunday';
+    expect(renderRecurrenceWeekdays(weekdays, interval)).toBe(expected);
+  });
+
+  it('should handle all weekdays in non-sequential order with an interval', () => {
+    const weekdays = ['wed', 'mon', 'fri', 'tue', 'sun', 'thu', 'sat'];
+    const interval = 5;
+    const expected =
+      'Every 5 weeks on Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, and Sunday';
+    expect(renderRecurrenceWeekdays(weekdays, interval)).toBe(expected);
+  });
+});
+
+describe('can render the recurrence weekdays with support for intervals whilst taking the locale into account', () => {
+  const unorderedWeekdays = ['wed', 'mon', 'fri', 'tue', 'sun', 'thu', 'sat'];
+  const interval = 3;
+  const cases = [
+    ['de', 'Montag, Dienstag, Mittwoch, Donnerstag, Freitag, Samstag und Sonntag'],
+    ['en-gb', 'Monday, Tuesday, Wednesday, Thursday, Friday, Saturday and Sunday'],
+    ['en-us', 'Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, and Sunday'],
+    ['fr', 'lundi, mardi, mercredi, jeudi, vendredi, samedi et dimanche'],
+    [
+      'pt-br',
+      'segunda-feira, terça-feira, quarta-feira, quinta-feira, sexta-feira, sábado e domingo',
+    ],
+    ['es', 'lunes, martes, miércoles, jueves, viernes, sábado y domingo'],
+    ['it', 'lunedì, martedì, mercoledì, giovedì, venerdì, sabato e domenica'],
+    ['tr', 'Pazartesi, Salı, Çarşamba, Perşembe, Cuma, Cumartesi ve Pazar'],
+    ['pl', 'poniedziałek, wtorek, środa, czwartek, piątek, sobota i niedziela'],
+    ['mn', 'Даваа, Мягмар, Лхагва, Пүрэв, Баасан, Бямба, Ням'],
+    ['uk', 'понеділок, вівторок, середа, четвер, п’ятниця, субота і неділя'],
+    ['hu', 'hétfő, kedd, szerda, csütörtök, péntek, szombat és vasárnap'],
+    ['zh-cn', '星期一、星期二、星期三、星期四、星期五、星期六和星期日'],
+  ];
+
+  test.each(cases)(
+    'should localize all weekdays in non-sequential order in %s with an interval',
+    (locale, expected) => {
+      const oldLocale = moment.locale();
+      moment.locale(locale);
+      const formatted = renderRecurrenceWeekdays(unorderedWeekdays, interval);
+      expect(formatted).toBe(`Every 3 weeks on ${expected}`);
       moment.locale(oldLocale);
     }
   );