+++ indico/modules/rb/client/js/components/TimeInformation.jsx
@@ -77,13 +77,10 @@ function TimeInformation({
         {recurrence.type === 'every' && recurrence.interval === 'week' && recurrenceWeekdays && (
           <Segment>
             <div>
-              <div>
+              <strong>
                 <Icon name="sync alternate" />
-                <Translate as="strong">
-                  Every{' '}
-                  <Param name="weekdays" value={renderRecurrenceWeekdays(recurrenceWeekdays)} />
-                </Translate>
-              </div>
+                {renderRecurrenceWeekdays(recurrenceWeekdays, recurrence.number)}
+              </strong>
             </div>
           </Segment>
         )}